+++ src/textual_code/widgets/main_view.py
@@ -540,6 +540,9 @@ class MainView(Static):
                     self.app, "default_syntax_theme", "monokai"
                 ),
                 default_word_wrap=getattr(self.app, "default_word_wrap", False),
+                default_show_indentation_guides=getattr(
+                    self.app, "default_show_indentation_guides", True
+                ),
                 default_warn_line_ending=getattr(
                     self.app, "default_warn_line_ending", True
                 ),